third_party/blink/web_tests/platform/generic/fast/dom/Document/missing-arguments-expected.txt

PASS document.adoptNode() threw exception TypeError: Failed to execute 'adoptNode' on 'Document': 1 argument required, but only 0 present..
PASS document.createAttribute() threw exception TypeError: Failed to execute 'createAttribute' on 'Document': 1 argument required, but only 0 present..
PASS document.createAttributeNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'createAttributeNS' on 'Document': 2 arguments required, but only 1 present..
PASS document.createAttributeNS() threw exception TypeError: Failed to execute 'createAttributeNS' on 'Document': 2 arguments required, but only 0 present..
PASS document.createCDATASection() threw exception TypeError: Failed to execute 'createCDATASection' on 'Document': 1 argument required, but only 0 present..
PASS document.createComment() threw exception TypeError: Failed to execute 'createComment' on 'Document': 1 argument required, but only 0 present..
PASS document.createElement() threw exception TypeError: Failed to execute 'createElement' on 'Document': 1 argument required, but only 0 present..
PASS document.createElementNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'createElementNS' on 'Document': 2 arguments required, but only 1 present..
PASS document.createElementNS() threw exception TypeError: Failed to execute 'createElementNS' on 'Document': 2 arguments required, but only 0 present..
PASS document.createEvent() threw exception TypeError: Failed to execute 'createEvent' on 'Document': 1 argument required, but only 0 present..
PASS document.createProcessingInstruction('xml') threw exception TypeError: Failed to execute 'createProcessingInstruction' on 'Document': 2 arguments required, but only 1 present..
PASS document.createProcessingInstruction() threw exception TypeError: Failed to execute 'createProcessingInstruction' on 'Document': 2 arguments required, but only 0 present..
PASS document.createTextNode() threw exception TypeError: Failed to execute 'createTextNode' on 'Document': 1 argument required, but only 0 present..
PASS document.elementFromPoint() threw exception TypeError: Failed to execute 'elementFromPoint' on 'Document': 2 arguments required, but only 0 present..
PASS document.elementFromPoint(0) threw exception TypeError: Failed to execute 'elementFromPoint' on 'Document': 2 arguments required, but only 1 present..
PASS document.execCommand() threw exception TypeError: Failed to execute 'execCommand' on 'Document': 1 argument required, but only 0 present..
PASS document.getElementById() threw exception TypeError: Failed to execute 'getElementById' on 'Document': 1 argument required, but only 0 present..
PASS document.getElementsByClassName() threw exception TypeError: Failed to execute 'getElementsByClassName' on 'Document': 1 argument required, but only 0 present..
PASS document.getElementsByName() threw exception TypeError: Failed to execute 'getElementsByName' on 'Document': 1 argument required, but only 0 present..
PASS document.getElementsByTagName() threw exception TypeError: Failed to execute 'getElementsByTagName' on 'Document': 1 argument required, but only 0 present..
PASS document.getElementsByTagNameNS('http://www.w3.org/2000/svg') threw exception TypeError: Failed to execute 'getElementsByTagNameNS' on 'Document': 2 arguments required, but only 1 present..
PASS document.getElementsByTagNameNS() threw exception TypeError: Failed to execute 'getElementsByTagNameNS' on 'Document': 2 arguments required, but only 0 present..
PASS document.importNode() threw exception TypeError: Failed to execute 'importNode' on 'Document': 1 argument required, but only 0 present..
PASS document.queryCommandEnabled() threw exception TypeError: Failed to execute 'queryCommandEnabled' on 'Document': 1 argument required, but only 0 present..
PASS document.queryCommandIndeterm() threw exception TypeError: Failed to execute 'queryCommandIndeterm' on 'Document': 1 argument required, but only 0 present..
PASS document.queryCommandState() threw exception TypeError: Failed to execute 'queryCommandState' on 'Document': 1 argument required, but only 0 present..
PASS document.queryCommandSupported() threw exception TypeError: Failed to execute 'queryCommandSupported' on 'Document': 1 argument required, but only 0 present..
PASS document.queryCommandValue() threw exception TypeError: Failed to execute 'queryCommandValue' on 'Document': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
